split coco and finetune datasets
