+++ datasets/coco_metadata/finetune.yaml
@@ -1,0 +1,7 @@
+path: finetune
+train: images/train
+val: images/val
+
+names:
+  0: person
+  1: cat
