+++ datasets/coco_metadata/finetune_crop.yaml
@@ -1,0 +1,7 @@
+path: finetune_crop
+train: images/train
+val: images/val
+
+names:
+  0: person
+  1: cat
