+++ datasets/coco_metadata/finetune_wide.yaml
@@ -1,0 +1,7 @@
+path: finetune_wide
+train: images/train
+val: images/val
+
+names:
+  0: person
+  1: cat
