+++ models/yolo26n_finetune_ncnn_model/metadata.yaml
@@ -1,0 +1,21 @@
+description: Ultralytics YOLO26n model trained on datasets/coco_metadata/finetune.yaml
+author: Ultralytics
+date: '2026-07-31T12:35:18.684041+01:00'
+version: 8.4.113
+license: AGPL-3.0 License (https://ultralytics.com/license)
+docs: https://docs.ultralytics.com
+stride: 32
+task: detect
+head: Detect
+batch: 1
+imgsz:
+- 640
+- 640
+names:
+  0: person
+  1: cat
+args:
+  batch: 1
+  quantize: 16
+channels: 3
+end2end: false
