--- models/yolo26n_subclass_ncnn_model/metadata.yaml
@@ -1,20 +1,0 @@
-description: Ultralytics YOLO26n model trained on datasets/coco_metadata/coco_subclass_downsample_finetune.yaml
-author: Ultralytics
-date: '2026-07-29T16:57:04.762030'
-version: 8.4.76
-license: AGPL-3.0 License (https://ultralytics.com/license)
-docs: https://docs.ultralytics.com
-stride: 32
-task: detect
-batch: 1
-imgsz:
-- 640
-- 640
-names:
-  0: person
-  1: cat
-args:
-  batch: 1
-  half: true
-channels: 3
-end2end: false
