+++ src/settings.toml
@@ -11,7 +11,7 @@ LOG_LEVEL = "INFO" # Logging verbosity
 MAX_DETS = 5 # Max detections per frame
 NMS_IOU_THRESHOLD = 0.3 # IoU threshold for NMS
 MJPEG_QV = 5 # MJPEG quality value
-MODEL_DETECTION_PATH = "yolo26n_subclass_ncnn_model" # Detection model location
+MODEL_DETECTION_PATH = "yolo26n_finetune_ncnn_model" # Detection model location
 MODEL_EMBEDDING_PATH = "shufflenetv2_x0_5_embeddings_onnx_model" # Embedding model location
 MONITORING_PERIOD = 3 # Seconds between monitoring checks
 OUTPUT_DIR = "object_clips" # Output folder for clips/logs
